--- a/CHANGELOG.md
+++ b/CHANGELOG.md
@@ -30,7 +30,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix failing PHP integration tests [2378](https://github.com/microsoft/kiota/issues/2378)
 - Prevents method overloading for go getters and setters with different values. [#2719](https://github.com/microsoft/kiota/issues/2719)
 - Fixed PHP request executor methods that return enums.
-- Allow configuration of the number of threads via the environment variable KIOTA_GENERATION_MDOP.
+- Allow configuration of the number of threads via the environment variable KIOTA_GENERATION_MAXDEGREEOFPARALLELISM.
 
 ## [1.3.0] - 2023-06-09