--- a/src/Kiota.Web/wwwroot/appsettings.json
+++ b/src/Kiota.Web/wwwroot/appsettings.json
@@ -1,6 +1,7 @@
 {
   "Generation": {
-    "IgnoredRequestContentTypes": ["multipart/form-data"]
+    "IgnoredRequestContentTypes": ["multipart/form-data"],
+    "MaxDegreeOfParallelism": -1
   },
   "Search": {
     "APIsGuruListUrl": "https://raw.githubusercontent.com/APIs-guru/openapi-directory/gh-pages/v2/list.json",