This causes quite severe corruption of the CodeModel. Many threads concurrently start creating the same models and start adding properties. The result is undefined for larger models and will be different on every run.

This probably fixes (or at least improves the situation for) #2442.

I ran into this trying to build an SDK with the fix for the renaming of properties, which sometimes worked and sometimes didn't work.