@papegaaij 2023-07-05
I really do not see any way to guarantee consistency when the CodeModel is built in a concurrent way. At the moment, multiple threads start creating the same CodeClasses and start building on those partly generated classes further on. Getting this to be thread safe will require a lot of explicit locking around building types, which will very likely remove any benefit from running this in parallel. For our API (about 10% the size of the Graph API), I didn't notice any change in performance. Generation is still done in just a few seconds. Generating the Go code does not take any longer than compiling it. IMHO the benefits of parallelism at this point do not outweigh the caveats of all the concurrency issues.