@papegaaij 2023-07-05
My simple benchmark on my gigantic machine :grin: (yes, it is and my colleagues think I'm mad carrying such a heavy 'laptop'), gave about the following results:
With parallelism: Total generation for Go takes a little over 30 seconds, of which about 11-12 is spent in step 7.
Without parallelism: Total generation takes about 60 seconds, of which 40 is spent in step 7.

So the speedup on my machine is roughly 30 seconds or 50%. But, the instability also counts. I had several runs with parallelism that crashed with an error. Given the unpredictability, I would blame the concurrency. Also, none of the runs with parallelism produced compiling code. The runs without parallelism did produce working code.

Compiling the Go code takes about 30 seconds. IMHO this must be added to the total build time. This reduces the speed gain to about 33%.