@papegaaij 2023-07-06
These are the proplems I've spotted for the creation of properties:
* KiotaBuilder.CreatePropertiesForModelClass can be run by multiple threads at the same time for the same type. This results CodeClass.AddProperty to start resolving conflicts for identical properties.
* CodeClass.AddProperty searches through the type hierarchy for existing properties, but has no guarantee that these types are actually built entirely. This often results in it not picking up properties, because the super type did exist, but its properties hadn't been created yet.

However, during my runs I've also seen problems building inheritance chains and I've got infos about unused types being pruned even though I'm sure they are being used in the API.

Building a complex, highly interconnected data structure concurrently really is asking for trouble if you ask me. I'd much rather have a tool that's highly reliable but a bit slower than a tool that's fast but spits out incorrect results every now and then. These concurrency issues will continue to haunt you and they will be very hard to find and nearly impossible to reproduce.

What's more, the concurrency causes code paths to be executed in unpredictable ways. This can cause Kiota to make different decisions depending on which happens first. Even if all concurrency issues get fixed, this can still cause Kiota to generate different code on subsequent runs on the same input. IMHO this should never happen in a tool like Kiota.